+++ NEWS.md
@@ -4,13 +4,14 @@ rdrawing 1.1£º
   - fix: ImageDrawSectorEx error
   - fix: ImageFillTriangleEx dead lock, when three vertices are on the same vertical line
   - fix: ImageDrawLineEx error, when length of line is shorter than line width
-  - enhancement: add setAutoUpdateWorld()/updateWorld()/waitMilliseconds() functions
 
 rturtle 1.1:
 
   - change: clear() don't reset turtle to it's origin
   - enhancement: add waitClose()
   - fix: truncation error not correctly handled in fd()
+  - add: beginFill() / endFill() / isFilling()
+  - enhancement: add setAutoUpdateWorld()/updateWorld()/waitMilliseconds() functions
 
 rturtle 1.0
 rdrawing 1.0